+++ test/index.js
@@ -52,6 +52,15 @@ testRule(1, { atRulesDontCount: true }, function(tr) {
   tr.notOk('a {{ @media print {{ b {{ c { top: 0; }}}}}}}', rule.messages.rejected);
 });
 
+testRule(1, { countNestedAtRules: false }, function(tr) {
+  basics(tr);
+
+  tr.ok('a { @media print { b { top: 0; }}}');
+  tr.ok('a { b { @media print { top: 0; }}}');
+  tr.notOk('a { b { @media print { c { top: 0; } }}}', rule.messages.rejected);
+  tr.notOk('a { @media print { b { c { top: 0; }}}}', rule.messages.rejected);
+});
+
 function basics(tr) {
   tr.ok('');
   tr.ok('a {}');