+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Head
+
+- Add `countNestedAtRules` option.
+
 ## 1.0.0
 
 - Remove PostCSS 4 compatibility.